Fetch and decode /Summary: the watch's own exact totals
/Summary is a third transport, distinct from /Data's bulk stream and
/Entries' single handle-fetch: GET, then repeated TYPE=0x0D reads at
increasing byte offsets, each reply a 19-byte header plus up to 451 payload
bytes, with a status at header offset 6 saying whether more pages follow
(100) or this was the last (200). The payloads concatenate into a plain
SBEM0103 container - no Heatshrink.

Inside is chunk 0x1b, the 133-field Header group that /Data never carries,
which is why ascent/descent/energy were never found there. Decoded against
the real captured Summary for the cycling fixture:

  Duration 4736.15s, PauseDuration 2436.33s -> moving 2299.8s
  Distance 5417m, Ascent 27.84m, Descent 19.47m
  Altitude 98.65-107.52m, Energy 651466 J = 155.7 kcal

Two independent checks land: moving time is exactly the 38:19 the app
reports (the GPS-gap heuristic gets -7%), and the altitude range matches
the delta-chain reconstruction added in the previous commit to ~0.1 m,
having been derived a completely different way.

Also fixes a latent bug: Sbem::parseContainer() didn't implement the
chunk-id 0xFF escape, so any id above 254 would have been misparsed.
Chunk::id widened to uint16_t. The descriptor table runs past id 300.

syncWatchWorkouts() now fetches /Summary after each /Data and overlays the
exact totals; a failed Summary fetch is non-fatal and the /Data-derived
workout is still saved. Not yet run on real hardware.

+++ src/ble/sbemcontainer.h
@@ -28,7 +28,12 @@ namespace Sbem {
 std::vector<uint8_t> heatshrinkDecompress(const std::vector<uint8_t> &compressed);
 
 struct Chunk {
-    uint8_t id;
+    // 16-bit, not 8: a chunk id of 0xFF means the real id is the next two
+    // bytes, little-endian (confirmed by decompiling libmds.so's
+    // BSML::SmlStreamParser::parseChunkHeader - the length field's own
+    // 0xFF escape has the same shape). Ids above 254 do occur: the
+    // descriptor table itself runs past 300.
+    uint16_t id;
     std::vector<uint8_t> value;
 };
